ocr_batch_tool/requirements.txt
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
PyPDF2==3.0.1
tqdm==4.66.1